requirements.txt
streamlit==1.26.0
websocket-client==1.8.0